MassBalanceOpt.jl

A package for formulating mass balance models of chemical processes and solving them with JuMP.

Flowsheets, Streams, and Blocks

MassBalanceOpt.AbstractBlock — Type

AbstractBlock

Abstract supertype for all blocks (Mixer, Splitter, etc.)

See also Mixer, Splitter, Separator, YieldReactor, MultiYieldReactor, StoicReactor, @Block_fields, @Block_init, @Block_finish

MassBalanceOpt.Flowsheet — Type

Flowsheet(name::Symbol=:index, parent=nothing)

Create a flowsheet. If name is omitted it defaults to :index. If specified, parent must be another Flowsheet; The new flowsheet will be a child of the parent flowsheet. If parent is omitted, the flowsheet has no parent.

A Flowsheet serves as a container for streams, blocks (subtypes of AbstractBlock), and other flowsheets. The names of a flowsheet's ancestors are embedded in the variable and equation names of the blocks and streams contained in the flowsheet. For example, if a flowsheet named :unit1 is a child of the index flowsheet (a flowsheet created with no parent), the variable and equation names in that flowsheet will start with unit1_. If the :unit1 flowsheet has a child named reactors the names will start with unit1_reactors_. Variables and equations in a flowsheet with no parent have no prefix.

Examples

Create a :index flowsheet with no parent:

julia> fs = Flowsheet()
Flowsheet(
   name=index   )

Add a Stream to the :index flowsheet:

julia> in1 = Stream(:in1, fs, @components(h2,ch4))
Stream(in1, fs=index, basis=FRAC, components=[h2, ch4])

julia> fs
Flowsheet(
   name=index,
   streams=[in1]   )

Create a flowsheet named :unit1 whose parent is :index:

julia> fs_1 = Flowsheet(:unit1, fs)
Flowsheet(
   name=unit1,
   parent=index   )

julia> fs
Flowsheet(
   name=index,
   children=[unit1],
   streams=[in1]   )

See also Stream, AbstractBlock

MassBalanceOpt.@components — Macro

@components symbols...

Create and return a component set made up of one or more unquoted symbols.

Examples

julia> comps1 = @components h2 ch4 c2h4
OrderedSet{Symbol} with 3 elements:
  :h2
  :ch4
  :c2h4

julia> comps1 = @components(h2, ch4, c2h4)
OrderedSet{Symbol} with 3 elements:
  :h2
  :ch4
  :c2h4

See also Stream

MassBalanceOpt.Stream — Type

Representation of a material stream in a mass balance model.

Stream(name::Symbol, fs, comps; basis=FRAC)

Create a new Stream with name and components comps in flowsheet fs. The keyword argument basis can be FRAC or FLOW, with FRAC being the default.

Creating a Stream does not create any variables or equations in the model. The stream becomes available for use as an argument to the constructors of blocks, which are subtyped of AbstractBlock. Blocks are responsible for creating stream variables and equations. When a block creates stream variables it uses the basis of each stream to determine what variables to create. If basis=FRAC a mass fraction variable is created for every component in the stream; if basis=FLOW a component mass flow rate variable is created for every component.

Warning

None of the built-in blocks create equations to sum the mass fractions or the component mass flow rates.

Examples

Create two inlet streams and one outlet stream, then create a Mixer block that mixes the two inlet streams.

julia> m = Model(); fs = Flowsheet();

julia> in1 = Stream(:in1, fs, @components(h2, ch4))
Stream(in1, fs=index, basis=FRAC, comps=[h2, ch4])

julia> in2 = Stream(:in2, fs, @components(h2, ch4, c2h4), basis=FLOW)
Stream(in2, fs=index, basis=FLOW, comps=[h2, ch4, c2h4])

julia> out = Stream(:out, fs, @components(h2, ch4, c2h4))
Stream(out, fs=index, basis=FRAC, comps=[h2, ch4, c2h4])

julia> mix1 = Mixer(m, :mix1, fs, [in1, in2], out)
Mixer(in=[in1, in2], out=[out])

julia> print_vars(mix1)
              Name              |    Value   |    Lower   |    Upper   |    Start   |Spec
--------------------------------|------------|------------|------------|------------|----
mix1_in1_mass                   |           1|            |            |            |F
mix1_in1_h2_massfrac            |         0.5|            |            |            |F
mix1_in1_ch4_massfrac           |         0.5|            |            |            |F
mix1_in2_mass                   |           1|            |            |            |F
mix1_in2_h2_mass                |     0.33333|            |            |            |F
mix1_in2_ch4_mass               |     0.33333|            |            |            |F
mix1_in2_c2h4_mass              |     0.33333|            |            |            |F
mix1_out_mass                   |            |            |            |            |
mix1_out_h2_massfrac            |            |            |            |            |
mix1_out_ch4_massfrac           |            |            |            |            |
mix1_out_c2h4_massfrac          |            |            |            |            |
11 variables

See also @stream, @streams, Flowsheet, @components, StreamBasis

MassBalanceOpt.StreamBasis — Type

@enum StreamBasis

FRAC: Generate mass fraction variables (default). FLOW: Generate component mass flow rate variables.

See also Stream

MassBalanceOpt.is_frac — Function

is_frac(strm::Stream)

Return true if the strm basis=FRAC, otherwise return false.

See also is_flow, StreamBasis

MassBalanceOpt.is_flow — Function

is_flow(strm::Stream)

Return true if the strm basis=FLOW, otherwise return false.

See also is_frac, StreamBasis

MassBalanceOpt.@stream — Macro

@stream(name, comps, ex_basis=:(basis=FRAC))

Create a Stream named name with component set comps and StreamBasis basis. @stream assumes that the model is stored in a variable named m and the active flowsheet is stored in a variable named fs in the current scope.

Examples

julia> feed = @stream(feed, comps_feed)

julia> prod = @stream(feed, comps_prod, basis=FLOW)

See also Stream, @streams, StreamBasis

MassBalanceOpt.@streams — Macro

@streams(blk)

Create a tuple of Streams using the expressions in the begin/end block blk. @streams assumes that the model is stored in a variable named m and the active flowsheet is stored in a variable named fs in the current scope.

Examples

julia> (ethfd, profd, cgcout) = @streams(begin
    ethfd, comps_ethfeed
    profd, comps_profeed
    cgcout, comps_cg, (basis=FLOW)
end)

See also Stream, @stream, StreamBasis

MassBalanceOpt.copy_stream — Function

copy_stream(strm::Stream)

Copy the values of the strm variables in strm's from block into the values of the strm variables in strm's to block. If either the to or from field of strm is empty, do nothing.

See also copy_streams

MassBalanceOpt.copy_streams — Function

copy_streams(strms::Vector{Stream})

Invoke copy_stream for each element of strms.

See also copy_stream

MassBalanceOpt.make_stream_vars! — Function

make_stream_vars!(m::Model, strm::Stream, prefix::AbstractString, var_list::Vector{VariableRef})
    -> Dict{Symbol, Any}

Create stream variables for strm using prefix at the beginning of the variable names. Add the new variables to var_list. Return the stream variables in a dictionary.

See also Stream, set_stream_var_specs!, make_var!

make_stream_vars!(m::Model, strms::Vector{Stream}, prefix::AbstractString, var_list::Vector{VariableRef})
    -> Dict{Symbol, Dict}

Create stream variables for all the streams in the array strms and return them in a dictionary.

MassBalanceOpt.set_stream_var_specs! — Function

set_stream_var_specs!(strms::Vector{Stream}, vars::Dict{Symbol, Dict})

Fix the variables stored in vars, for each stream in strms, equal to default values.

See also Stream, make_stream_vars!, fix

MassBalanceOpt.Mixer — Type

Mixer <: AbstractBlock

Mix two or more inlet streams into one outlet stream.

Create a Mixer block:

    Mixer(m::Model, name::Symbol, fs::Flowsheet, inlets::Vector{Stream}, outlet::Stream)

or if the current scope contains m and fs bound to a Model and Flowsheet:

    @block(name, Mixer, inlets, outlet)

Examples

julia> m = Model(); fs = Flowsheet(); comps = @components A B;

julia> (in1, in2, out) = @streams begin
           in1, comps
           in2, comps
           out, comps
       end;

julia> mix1 = @block(mix1, Mixer, [in1, in2], out);

julia> print_vars(mix1)
              Name               Fix      Value          Lower          Upper          Start     
--------------------------------|---|--------------|--------------|--------------|--------------|
mix1_in1_mass                     ==              1|              |              |             1|
mix1_in1_A_massfrac               ==            0.5|              |              |           0.5|
mix1_in1_B_massfrac               ==            0.5|              |              |           0.5|
mix1_in2_mass                     ==              1|              |              |             1|
mix1_in2_A_massfrac               ==            0.5|              |              |           0.5|
mix1_in2_B_massfrac               ==            0.5|              |              |           0.5|
mix1_out_mass                                      |              |              |              |
mix1_out_A_massfrac                                |              |              |              |
mix1_out_B_massfrac                                |              |              |              |
9 variables

See also Stream, Splitter, Separator, YieldReactor, MultiYieldReactor, StoicReactor

MassBalanceOpt.Splitter — Type

Splitter <: AbstractBlock

Split one inlet stream into two or more outlet streams.

Create a Splitter block:

    Splitter(m::Model, name::Symbol, fs::Flowsheet, inlet::Stream, outlets::Vector{Stream})

or if the current scope contains m and fs bound to a Model and Flowsheet:

    @block(name, Splitter, inlet, outlets)

Examples

julia> m = Model(); fs = Flowsheet(); comps = @components A B;

julia> (in1, out1, out2) = @streams begin
           in1, comps
           out1, comps
           out2, comps
       end;

julia> spl = @block(spl, Splitter, in1, [out1, out2]);

julia> print_vars(spl)
              Name               Fix      Value          Lower          Upper          Start     
--------------------------------|---|--------------|--------------|--------------|--------------|
spl_in1_mass                      ==              1|              |              |             1|
spl_in1_A_massfrac                ==            0.5|              |              |           0.5|
spl_in1_B_massfrac                ==            0.5|              |              |           0.5|
spl_out1_mass                                      |              |              |              |
spl_out1_A_massfrac                                |              |              |              |
spl_out1_B_massfrac                                |              |              |              |
spl_out2_mass                                      |              |              |              |
spl_out2_A_massfrac                                |              |              |              |
spl_out2_B_massfrac                                |              |              |              |
spl_out1_split_frac               ==            0.5|              |              |           0.5|
spl_out2_split_frac                                |              |              |           0.5|
11 variables

See also Stream, Mixer, Separator, YieldReactor, MultiYieldReactor, StoicReactor

MassBalanceOpt.Separator — Type

Separator <: AbstractBlock

Separate the components in the inlet stream into two or more outlet streams.

Create a Separator block:

    Separator(m::Model, name::Symbol, fs::Flowsheet, inlet::Stream, outlets::Vector{Stream})

or if the current scope contains m and fs bound to a Model and Flowsheet:

    @block(name, Separator, inlet, outlets)

Examples

julia> m = Model(); fs = Flowsheet(); comps = @components A B;

julia> (in1, out1, out2) = @streams begin
    in1, comps
    out1, comps
    out2, comps
end;

julia> sep = @block(sep, Separator, in1, [out1, out2]);

julia> @set sep_A_out1_split = 0.3

julia> @set sep_B_out1_split = 0.6

julia> set_start_values(sep)

julia> print_vars(sep)
              Name               Fix      Value          Lower          Upper          Start     
--------------------------------|---|--------------|--------------|--------------|--------------|
sep_in1_mass                      ==              1|              |              |             1|
sep_in1_A_massfrac                ==            0.5|              |              |           0.5|
sep_in1_B_massfrac                ==            0.5|              |              |           0.5|
sep_out1_mass                                      |              |              |          0.45|
sep_out1_A_massfrac                                |              |              |     0.3333333|
sep_out1_B_massfrac                                |              |              |     0.6666667|
sep_out2_mass                                      |              |              |          0.55|
sep_out2_A_massfrac                                |              |              |     0.6363636|
sep_out2_B_massfrac                                |              |              |     0.3636364|
sep_in1_A_mass                                     |              |              |           0.5|
sep_in1_B_mass                                     |              |              |           0.5|
sep_out1_A_mass                                    |              |              |          0.15|
sep_out1_B_mass                                    |              |              |           0.3|
sep_out2_A_mass                                    |              |              |          0.35|
sep_out2_B_mass                                    |              |              |           0.2|
sep_A_out1_split                  ==            0.3|              |              |           0.3|
sep_A_out2_split                                   |              |              |           0.7|
sep_B_out1_split                  ==            0.6|              |              |           0.6|
sep_B_out2_split                                   |              |              |           0.4|
19 variables

See also Stream, Mixer, Splitter, YieldReactor, MultiYieldReactor, StoicReactor

MassBalanceOpt.@stoic — Macro

@stoic(expr)

Create an array of OrderedDicts, each of which encodes the stoichiometric coefficients of a single reaction. The expr should be one or more balanced chemical reactions written in the form:

aA + bB + ... => cC + dD

where a, b, etc. are real or integer stoichiometric coefficients, and A, B, etc. are species. The returned array looks like:

coef[1] = OrderedDict(:A => -a, :B => -b, :C => c, :D => d)   # coefficients for reaction 1
coef[2] = OrderedDict(:A => -a, :B => -b, :C => c, :D => d)   # coefficients for reaction 2
  ...

Note that the coefficients in the OrderedDict are positive for products and negative for reactants.

Examples

julia> coef = @stoic A + B => C
julia> coef
1-element Vector{OrderedCollections.OrderedDict{Symbol, Real}}:
 OrderedCollections.OrderedDict(:A => -1, :B => -1, :C => 1)

julia> @stoic begin
        A + B => C
      2A + 3B => 4C + D
    end
2-element Vector{OrderedCollections.OrderedDict{Symbol, Real}}:
OrderedCollections.OrderedDict(:A => -1, :B => -1, :C => 1)
OrderedCollections.OrderedDict(:A => -2, :B => -3, :C => 4, :D => 1)

See also StoicReactor

MassBalanceOpt.StoicReactor — Type

StoicReactor <: AbstractBlock

Model a reactor in which a set of stoichiometric reactions with specified conversions takes place.

Create a StoicReactor block:

    StoicReactor(m::Model, name::Symbol, fs::Flowsheet, inlet::Stream, outlet::Stream,
        stoic_coef::Vector{OrderedDict{Symbol, Real}}, 
        mw::Dict{Symbol, Float64},
        conv::OrderedDict{Int, NamedTuple{(:c, :X), Tuple{Symbol, Float64}}}= OrderedDict{Int, NamedTuple{(:c, :X),
                               Tuple{Symbol, Float64}}}())

or if the current scope contains m and fs bound to a Model and Flowsheet:

    @block(name, StoicReactor, inlet, outlet, stoic_coef, mw, conv)

Examples

julia> m = Model(); fs = Flowsheet(); comps1 = @components A B; comps2 = @components A B C D;

julia> (in1, out1) = @streams begin
    in1, comps1
    out1, comps2
end;

julia> mw = Dict(:A => 30.0, :B => 28.0, :C => 35.0, :D => 30.0);

julia> coef = @stoic A + B => C + D
1-element Vector{OrderedDict{Symbol, Real}}:
 OrderedDict(:A => -1, :B => -1, :C => 1, :D => 1)

 julia> conv = OrderedDict(1 => (c=:A, X=0.65))
 OrderedDict{Int64, NamedTuple{(:c, :X), Tuple{Symbol, Float64}}} with 1 entry:
   1 => (c = :A, X = 0.65)
   
julia> r1 = @block(r1, StoicReactor, in1, out1, coef, mw, conv);

julia> print_vars(r1)
              Name               Fix      Value          Lower          Upper          Start     
--------------------------------|---|--------------|--------------|--------------|--------------|
r1_in1_mass                       ==              1|              |              |             1|
r1_in1_A_massfrac                 ==            0.5|              |              |           0.5|
r1_in1_B_massfrac                 ==            0.5|              |              |           0.5|
r1_out1_mass                                       |              |              |              |
r1_out1_A_massfrac                                 |              |              |              |
r1_out1_B_massfrac                                 |              |              |              |
r1_out1_C_massfrac                                 |              |              |              |
r1_out1_D_massfrac                                 |              |              |              |
r1_in1_A_mass                                      |              |              |              |
r1_in1_B_mass                                      |              |              |              |
r1_out1_A_mass                                     |              |              |              |
r1_out1_B_mass                                     |              |              |              |
r1_out1_C_mass                                     |              |              |              |
r1_out1_D_mass                                     |              |              |              |
r1_in1_A_moles                                     |              |              |              |
r1_in1_B_moles                                     |              |              |              |
r1_out1_A_moles                                    |              |              |              |
r1_out1_B_moles                                    |              |              |              |
r1_out1_C_moles                                    |              |              |              |
r1_out1_D_moles                                    |              |              |              |
r1_extent_rx_1                                     |              |              |              |
r1_conv_A_rx_1                    ==           0.65|              |              |          0.65|
22 variables

See also @stoic, Stream, YieldReactor, MultiYieldReactor

MassBalanceOpt.YieldReactor — Type

YieldReactor <: AbstractBlock

Model a reactor with one inlet and one outlet stream, in which each component in the inlet stream has a specified yield to the components in the outlet stream.

Create a YieldReactor block:

    YieldReactor(m::Model, name::Symbol, fs::Flowsheet, inlet::Stream, outlet::Stream)

julia or if the current scope contains m and fs bound to a Model and Flowsheet:

    @block(name, YieldReactor, inlet, outlet)

Examples

julia> m = Model(); fs = Flowsheet(); comps1 = @components A B; comps2 = @components A B C D;

julia> (in1, out1) = @streams begin
    in1, comps1
    out1, comps2
end;

julia> r1 = @block(r1, YieldReactor, in1, out1);

julia> @set begin
    r1_y_C_from_A = 0.3
    r1_y_D_from_A = 0.4
    r1_y_C_from_B = 0.6
    r1_y_D_from_B = 0.1
end

julia> set_start_values(r1)

julia> print_vars(r1)
Name                             Fix      Value          Lower          Upper          Start     
--------------------------------|---|--------------|--------------|--------------|--------------|
r1_in1_mass                       ==              1|              |              |             1|
r1_in1_A_massfrac                 ==            0.5|              |              |           0.5|
r1_in1_B_massfrac                 ==            0.5|              |              |           0.5|
r1_out1_mass                                       |              |              |             1|
r1_out1_A_massfrac                                 |              |              |          0.15|
r1_out1_B_massfrac                                 |              |              |          0.15|
r1_out1_C_massfrac                                 |              |              |          0.45|
r1_out1_D_massfrac                                 |              |              |          0.25|
r1_in1_A_mass                                      |              |              |           0.5|
r1_in1_B_mass                                      |              |              |           0.5|
r1_out1_A_mass                                     |              |              |          0.15|
r1_out1_B_mass                                     |              |              |          0.15|
r1_out1_C_mass                                     |              |              |          0.45|
r1_out1_D_mass                                     |              |              |          0.25|
r1_y_A_from_A                                      |              |              |           0.3|
r1_y_B_from_A                     ==              0|              |              |             0|
r1_y_C_from_A                     ==            0.3|              |              |           0.3|
r1_y_D_from_A                     ==            0.4|              |              |           0.4|
r1_y_A_from_B                     ==              0|              |              |             0|
r1_y_B_from_B                                      |              |              |           0.3|
r1_y_C_from_B                     ==            0.6|              |              |           0.6|
r1_y_D_from_B                     ==            0.1|              |              |           0.1|
22 variables

See also Stream, MultiYieldReactor, StoicReactor

MassBalanceOpt.MultiYieldReactor — Type

MultiYieldReactor <: AbstractBlock

Model a reactor with multiple inlet and outlet streams. This is used to represent a set of YieldReactors operating in parallel. The number of inlet streams must equal the number of outlet streams, and the length of feed_names must equal the number of streams.

Create a MultiYieldReactor block:

    MultiYieldReactor(m::Model, name::Symbol, fs::Flowsheet, inlets::Vector{Stream}, outlets::Vector{Stream},
         feed_names::Vector{Symbol}, reactor_name::Symbol)

or if the current scope contains m and fs bound to a Model and Flowsheet:

    @block(name, MultiYieldReactor, inlets, outlets, feed_names, reactor_name)

Examples

julia> m = Model(); fs = Flowsheet();

julia> compsA = @components(A); compsB = @components(B); comps_out = @components A B C D;

julia> (feedA, feedB, outA, outB) = @streams begin
    feedA, compsA
    feedB, compsB
    outA , comps_out
    outB , comps_out
end;

julia> r1 = @block(r1, MultiYieldReactor, [feedA, feedB], [outA, outB], [:feedA, :feedB], :rx);

julia> print_vars(r1)
              Name               Fix      Value          Lower          Upper          Start     
--------------------------------|---|--------------|--------------|--------------|--------------|
r1_feedA_mass                     ==              1|              |              |             1|
r1_feedA_A_massfrac               ==              1|              |              |             1|
r1_feedB_mass                     ==              1|              |              |             1|
r1_feedB_B_massfrac               ==              1|              |              |             1|
r1_outA_mass                                       |              |              |              |
r1_outA_A_massfrac                                 |              |              |              |
r1_outA_B_massfrac                                 |              |              |              |
r1_outA_C_massfrac                                 |              |              |              |
r1_outA_D_massfrac                                 |              |              |              |
r1_outB_mass                                       |              |              |              |
r1_outB_A_massfrac                                 |              |              |              |
r1_outB_B_massfrac                                 |              |              |              |
r1_outB_C_massfrac                                 |              |              |              |
r1_outB_D_massfrac                                 |              |              |              |
r1_feedA_A_mass                                    |              |              |              |
r1_feedB_B_mass                                    |              |              |              |
r1_outA_A_mass                                     |              |              |              |
r1_outA_B_mass                                     |              |              |              |
r1_outA_C_mass                                     |              |              |              |
r1_outA_D_mass                                     |              |              |              |
r1_outB_A_mass                                     |              |              |              |
r1_outB_B_mass                                     |              |              |              |
r1_outB_C_mass                                     |              |              |              |
r1_outB_D_mass                                     |              |              |              |
r1_total_feed_mass                                 |              |              |              |
r1_feedA_n_rx                                      |              |              |              |
r1_feedA_rate                     ==              1|              |              |             1|
r1_feedB_n_rx                                      |              |              |              |
r1_feedB_rate                     ==              1|              |              |             1|
r1_n_rx                                            |              |              |              |
r1_feedA_y_A_from_A                                |              |              |             0|
r1_feedA_y_B_from_A               ==              0|              |              |             0|
r1_feedA_y_C_from_A               ==              0|              |              |             0|
r1_feedA_y_D_from_A               ==              0|              |              |             0|
r1_feedB_y_A_from_B               ==              0|              |              |             0|
r1_feedB_y_B_from_B                                |              |              |             0|
r1_feedB_y_C_from_B               ==              0|              |              |             0|
r1_feedB_y_D_from_B               ==              0|              |              |             0|
38 variables

See also Stream, YieldReactor, StoicReactor

MassBalanceOpt.@block — Macro

@block(name, kind, args...)

Create a block named name with type kind in flowsheet fs using arguments args.... The value of kind must be a subtype of AbstractBlock. @block assumes that the model is stored in a variable named m and the active flowsheet is stored in a variable named fs in the current scope.

Examples

julia> ctdc1 = @block(ctdc1, Separator, arxo, [offgas, dc1bt])

julia> arx = @block(arx, StoicReactor, dc2oh, arxo, arx_stoic, arx_mw, arx_conv)

See also Mixer, Splitter, Separator, YieldReactor, MultiYieldReactor, StoicReactor

MassBalanceOpt.@Block_fields — Macro

@Block_fields

Create the fields typically used in a block. They are:

    name::Symbol
    fs::Flowsheet
    inlets::Vector{Stream}
    outlets::Vector{Stream}
    strm_vars::Dict{Symbol, Dict}
    var_list::Vector{VariableRef}
    eq_list::Vector{ConstraintRef}

See also AbstractBlock, @Block_init, @Block_finish

MassBalanceOpt.@Block_init — Macro

@Block_init

Create the stream variables used in a block and set their specs.

See also AbstractBlock, @Block_fields, @Block_finish

MassBalanceOpt.@Block_finish — Macro

@Block_finish

Assign a newly created block referenced by self to its container flowsheet, and record the block in the to and from fields of the inlet and outlet streams.

See also AbstractBlock, @Block_fields, @Block_init

MassBalanceOpt.make_mass_flow_vars! — Function

make_mass_flow_vars!(m::Model,
                     prefix::AbstractString,
                     inlets::Vector{Stream},
                     outlets::Vector{Stream},
                     strm_vars::Dict{Symbol, Dict},
                     var_list::Vector{VariableRef},
                     eq_list::Vector{ConstraintRef})::Dict{Symbol, Dict}

Conditionally create component mass flow rate variables for the streams in inlets and outlets. If the stream has a FLOW basis, component mass flow rate variables already exist. If the stream has a FRAC basis, make new component mass flow variables and equations to calculate them. Add equations to calculate the outlet stream total mass flow rates from the component mass flow rates.

JuMP.set_start_values — Method

set_start_values(blk::AbstractBlock; copy_inlets::Bool=true)

Set the start values of the block variables in blk. If copy_inlets=false, don't copy the inlet stream variable start values from the upstream block.

Variables and Equations

MassBalanceOpt.make_var! — Function

make_var!(m::Model, name::AbstractString, var_list::Vector{VariableRef})

Create a new JuMP variable named name, register it in model m as m[:name], and add it to var_list.

JuMP.fix — Method

fix(var::VariableRef, [val::Real])

If val is passed, fix the variable var equal to val and set its start value to val. If no val is passed, fix the variable equal to its value in the most recent solution. If no solution is available, fix the variable equal to its start value. If no start value is available, fix the variable equal to 0.

See also free, @specs

MassBalanceOpt.free — Function

free(var::VariableRef)

Delete the fixing constraint for variable var.

See also fix, @specs

MassBalanceOpt.flip — Function

flip(var1::VariableRef, [var2::VariableRef])

If var2 is passed, invert the specs of var1 and var2, i.e., if var1 is fixed and var2 is free, make var1 free and var2 fixed; if var1 is free and var2 is fixed, make var1 fixed and var2 free. If both vars have the same spec, do nothing.

If var2 is not passed, invert the spec of var1.

Warning

The newly fixed variable will have its value set to:

1. its value in the most recent solution
2. its start value
3. 0

in that order. Use set_value or @set to change the value of the newly fixed variable.

See also @set, @specs

MassBalanceOpt.connect — Function

connect(var_lhs::VariableRef, var_rhs::VariableRef)

Given two variables var_lhs and var_rhs, at least one of which is fixed, add a new constraint:

var_lhs == var_rhs

If both variables are fixed, free var_rhs. Otherwise free the single fixed variable. If both variables are free, don't add a new constraint and return nothing.

Return the newly created ConstraintRef. See also @set

connect(strm::Stream)

Connect all the variables in the source and destination blocks of strm. If the stream has no source, or no destination, do nothing.

connect(m::Model, strms::Vector{Stream})

Connect all the streams in the array strms.

connect(fs::Flowsheet)

Connect all the streams in the flowsheet fs.

JuMP.set_value — Method

set_value(var::VariableRef, val::Real)

If var is fixed, reset its fixed value to val. If var is free, set its start value to val.

See also fix, @set, @values

MassBalanceOpt.set_lower — Function

set_lower(var::VariableRef, lower::Real)

If var is not fixed, set its lower bound to lower.

See also set_upper, delete_lower

MassBalanceOpt.set_upper — Function

set_upper(var::VariableRef, lower::Real)

If var is not fixed, set its upper bound to upper.

See also set_lower, delete_upper

MassBalanceOpt.delete_lower — Function

delete_lower(var::VariableRef)

If var is not fixed and has a lower bound, delete its lower bound.

See also set_lower, delete_upper

MassBalanceOpt.delete_upper — Function

delete_upper(var::VariableRef)

If var is not fixed and has an upper bound, delete its upper bound.

See also set_upper, delete_lower

MassBalanceOpt.@set — Macro

@set expr

Do the operations described by expr (see examples below). @set assumes that the model is stored in a variable named m in the current scope.

Examples

Set the value of the variable var to 100.0. If the variable is fixed, set the fixed value to 100.0, otherwise set the start value to 100.0:

julia> @set var = 100.0

Connect variables var1 and var2. This adds the equation var1 == var2 to the model. At least one of the variables must be fixed:

julia> @set var1 = var2

Set the upper bound on var to 100.0:

julia> @set var <= 100.0

Set the upper bound on var to 100.0:

julia> @set var <= 100.0

Delete the upper bound on var:

julia> @set var <= Inf

Delete the lower bound on var:

julia> @set -Inf <= var

Delete the lower and upper bounds on var:

julia> @set -Inf <= var <= Inf

Flip the specs on var1 and var2. If var1 is fixed and var2 is free, free var1 and fix var2. If var1 is free and var2 is fixed, fix var1 and free var2. If both variables have the same spec, do nothing:

julia> @set var1 ~ var2

Flip the spec on var. If var is fixed, free var. If var is free, fix var:

julia> @set ~var

Fix var:

julia> @set +var

Free var:

julia> @set -var

Set the lower and upper bounds on var:

julia> @set 1.0 < var < 2.0
julia> @set 2.0 > var > 1.0

Combine several operations in a begin/end block:

julia> @set begin
    x > 1.0
    y < 2.0
    1.0 < z < 2.0
    p ~ q
end

See also @values, @specs, @bounds

MassBalanceOpt.@values — Macro

@values expr

Alias for @set.

See also @set, @specs, @bounds

MassBalanceOpt.@specs — Macro

@specs expr

Alias for @set.

See also @set, @values, @bounds

MassBalanceOpt.@bounds — Macro

@bounds expr

Alias for @set.

See also @set, @values, @specs

MassBalanceOpt.make_eq! — Function

make_eq!(m::Model, name::AbstractString, con::ConstraintRef, eq_list::Vector{ConstraintRef})

Register the equation con in model m using name as the base name. Add the equation to eq_list.

See also make_var!

Printing and Output

MassBalanceOpt.print_vars — Function

print_vars([io::IO], vars::Vector{VariableRef})

Print the variables in vars in a table.

See also print_fixed, print_free, print_bounds, print_active, print_eqs, write_vars

print_vars([io::IO], var::VariableRef)

Print a single variable var.

print_vars([io::IO], m::Model)

Print all the variables in model m.

print_vars([io::IO], blk::AbstractBlock)

Print all the variables in block blk.

print_vars([io::IO], m::Model, glob::AbstractString)

Print the variables in model m that match glob, where glob is a string containing one or more * wildcard characters.

Examples

julia> print_vars(m, "*mass")

julia> print_vars(m, "*cgcfeed*")

print_vars([io::IO], m::Model, pattern::Regex)

Print the variables in model m that match the regular expression pattern.

Examples

julia> print_vars(m, r"^.*mass$")

MassBalanceOpt.write_vars — Function

write_vars([io::IO], m::Model)

Write the values of all the variables in m to io in the format:

@values begin
    var_1 = 1.0
    var_2 = 2.0
     ...
end

MassBalanceOpt.print_fixed — Function

print_fixed([io::IO], m::Model)

Print all the fixed variables in model m.

See also print_free, print_bounds, print_active

print_fixed([io::IO], blk::AbstractBlock)

Print all the fixed variables in block blk.

MassBalanceOpt.print_free — Function

print_free([io::IO], m::Model)

Print all the free variables in model m.

See also print_fixed, print_bounds, print_active

print_free([op::IO], blk::AbstractBlock)

Print all the free variables in block blk.

MassBalanceOpt.print_bounds — Function

print_bounds([io::IO], m::Model)

Print the variables in m with lower or upper bounds.

See also print_fixed, print_free, print_active

print_bounds([io::IO], blk::AbstractBlock)

Print the variables in blk with lower or upper bounds.

MassBalanceOpt.print_active — Function

print_active([io::IO], m::Model)

Print the variables in m whose values equal their lower or upper bounds.

See also print_fixed, print_free, print_bounds

MassBalanceOpt.print_eqs — Function

print_eqs([io::IO], m::Model)

Print all the equations in model m.

See also print_vars

print_eqs([io::IO], blk::AbstractBlock)

Print all the equations in block blk.

print_eqs([io::IO], m::Model, eq::Symbol)

Print the equation registered in model m as m[:eq]

MassBalanceOpt.print_model — Function

print_model([io::IO], model_or_block::Union{Model, AbstractBlock})

Print all the variables and equations in a model or block specified by model_or_block.

See also print_vars, print_eqs

Solving Models

MassBalanceOpt.@solve — Macro

@solve

Run the function optimize!(m) where m is a JuMP model in the current scope.

MassBalanceOpt.eval_obj — Function

eval_obj(m::Model)

Evalulate the objective function in m, using the solution values if a solution exists, otherwise using the start values.

Index

MassBalanceOpt.AbstractBlock

MassBalanceOpt.Flowsheet

MassBalanceOpt.Mixer

MassBalanceOpt.MultiYieldReactor

MassBalanceOpt.Separator

MassBalanceOpt.Splitter

MassBalanceOpt.StoicReactor

MassBalanceOpt.Stream

MassBalanceOpt.StreamBasis

MassBalanceOpt.YieldReactor

JuMP.fix

JuMP.set_start_values

JuMP.set_value

MassBalanceOpt.connect

MassBalanceOpt.copy_stream

MassBalanceOpt.copy_streams

MassBalanceOpt.delete_lower

MassBalanceOpt.delete_upper

MassBalanceOpt.eval_obj

MassBalanceOpt.flip

MassBalanceOpt.free

MassBalanceOpt.is_flow

MassBalanceOpt.is_frac

MassBalanceOpt.make_eq!

MassBalanceOpt.make_mass_flow_vars!

MassBalanceOpt.make_stream_vars!

MassBalanceOpt.make_var!

MassBalanceOpt.print_active

MassBalanceOpt.print_bounds

MassBalanceOpt.print_eqs

MassBalanceOpt.print_fixed

MassBalanceOpt.print_free

MassBalanceOpt.print_model

MassBalanceOpt.print_vars

MassBalanceOpt.set_lower

MassBalanceOpt.set_stream_var_specs!

MassBalanceOpt.set_upper

MassBalanceOpt.write_vars

MassBalanceOpt.@Block_fields

MassBalanceOpt.@Block_finish

MassBalanceOpt.@Block_init

MassBalanceOpt.@block

MassBalanceOpt.@bounds

MassBalanceOpt.@components

MassBalanceOpt.@set

MassBalanceOpt.@solve

MassBalanceOpt.@specs

MassBalanceOpt.@stoic

MassBalanceOpt.@stream

MassBalanceOpt.@streams

MassBalanceOpt.@values